+++ src/feature/posts-filter/index.ts
@@ -1,0 +1,1 @@
+export { PostsFilter } from "./ui/PostsFilter"
